# ENH: Specialize some np.char functions to operate on raw C data