--- a/numpy/core/src/multiarray/stringlib/stringlib.h
+++ b/numpy/core/src/multiarray/stringlib/stringlib.h
@@ -1,0 +1,3 @@
+#define STRINGLIB(F)                stringlib_##F
+#define STRINGLIB_CHAR              char
+#define STRINGLIB_CHAR_ISALPHA      isalpha