--- a/numpy/core/src/multiarray/stringlib/ucs4lib.h
+++ b/numpy/core/src/multiarray/stringlib/ucs4lib.h
@@ -1,0 +1,4 @@
+#define STRINGLIB(F)                ucs4lib_##F
+#define STRINGLIB_CHAR              npy_ucs4
+#define STRINGLIB_CHAR_ISALPHA      Py_UNICODE_ISALPHA
+