--- a/numpy/core/src/multiarray/vec_string.h
+++ b/numpy/core/src/multiarray/vec_string.h
@@ -1,0 +1,7 @@
+#ifndef NUMPY_CORE_SRC_MULTIARRAY_VEC_STRING_H_
+#define NUMPY_CORE_SRC_MULTIARRAY_VEC_STRING_H_
+
+NPY_NO_EXPORT PyObject *
+_vec_string(PyObject *, PyObject *, PyObject *);
+
+#endif