… doc/release/upcoming_changes/24756.performance.rst
… numpy/core/src/multiarray/multiarraymodule.c
… numpy/core/src/multiarray/stringlib/README.txt
… numpy/core/src/multiarray/stringlib/undef.h