@mattip 2023-09-20
`np.char.xxx` functions operate on byte- or unicode- ndarrays. They iterate across the items in the array, applying the `xxx` operation on each item. The current code converts each item to a python `bytes` or `str` object, performs the operation, and then writes the contents back to the result array item. It should be faster to avoid the conversion and operate directly on the byte array/usc32 string. The technique here sets up an array of fast functions, and will use the fast alternative if one exists.

@lysnikolaou to prove the performance improves, could you run the appropriate `spin` command to generate a comparison (hopefully some `np.char` benchmarks already exist)?