@lysnikolaou 2023-09-20
Here's a couple of benchmarks I wrote (unfortunately there aren't any right now).

```python3
class NumPyChar(Benchmark):
    def setup(self):
        self.A = np.array([100*'x', 100*'y'])
        self.B = np.array([100*'z', 100*'p'])

    def time_add(self):
        np.char.add(self.A, self.B)

    def time_isalpha(self):
        np.char.isalpha(self.A)
```

And the results:

```
All benchmarks:

| Change   | Before [70834eb3] <main>   | After [ad0ceebb] <vec-string-optimization>   |   Ratio | Benchmark (Parameter)             |
|----------|----------------------------|----------------------------------------------|---------|-----------------------------------|
| -        | 17.6±0.5μs                 | 9.09±0.01μs                                  |    0.52 | bench_core.NumPyChar.time_add     |
| -        | 7.47±0.09μs                | 2.97±0.01μs                                  |    0.4  | bench_core.NumPyChar.time_isalpha |
```