@ngoldbaum 2023-09-20
Cool! Have you checked if this works with [stringdtype](https://github.com/numpy/numpy-user-dtypes/tree/main/stringdtype)? There are tests in that repo to make sure all the ufunc-likes in `np.char` works.

I have some vague ideas about making these a special kind of ufunc but if we can get some big improvements without doing that, even better.