@mattip 2023-09-20
Would it be possible to avoid converting the method output to a `PyObject*` which is then assigned to the output ndarray? After all you have access to the raw data of the output just like you have access to the raw data of the input.