@lysnikolaou 2023-09-21
> Would it be possible to avoid converting the method output to a `PyObject*` which is then assigned to the output ndarray?

Done in https://github.com/numpy/numpy/pull/24756/commits/462191b3bf455588807d2b4c15a2326f3a295409.