@mattip 2023-09-21
Could you add your benchmarks to benchmarks/benchmarks and run them again?