@lysnikolaou 2023-09-21
> Could you add your benchmarks to benchmarks/benchmarks and run them again?

Should I push these to this PR as well?