@lysnikolaou 2023-09-21
And here's the results with bigger lists!

Code:
```python3
class NumPyChar(Benchmark):
    def setup(self):
        self.A = np.array([100*'x', 100*'y'])
        self.B = np.array([100*'z', 100*'p'])
        self.C = np.array(1000 * ['aa'])
        self.D = np.array(1000 * ['bb'])

    def time_add_small_list_big_string(self):
        np.char.add(self.A, self.B)

    def time_isalpha_small_list_big_string(self):
        np.char.isalpha(self.A)

    def time_add_big_list_small_string(self):
        np.char.add(self.C, self.D)

    def time_isalpha_big_list_small_string(self):
        np.char.isalpha(self.C)
```

Results:
```
| Change   | Before [70834eb3] <main>   | After [462191b3] <vec-string-optimization>   |   Ratio | Benchmark (Parameter)                                   |
|----------|----------------------------|----------------------------------------------|---------|---------------------------------------------------------|
| -        | 7.70±0.05μs                | 2.90±0.06μs                                  |    0.38 | bench_core.NumPyChar.time_isalpha_small_list_big_string |
| -        | 18.0±0.2μs                 | 6.67±0.1μs                                   |    0.37 | bench_core.NumPyChar.time_add_small_list_big_string     |
| -        | 1.35±0.01ms                | 13.5±0.09μs                                  |    0.01 | bench_core.NumPyChar.time_isalpha_big_list_small_string |
| -        | 3.39±0.06ms                | 16.2±0.3μs                                   |    0    | bench_core.NumPyChar.time_add_big_list_small_string     |
```