@mattip 2023-09-21
If string dtype seems hard, we can do that in a follow-on PR. 

> Should I push these to this PR as well?

Yes please. Those results seem very encouraging. Could you add a release note?