@ngoldbaum 2023-09-21
Is there a rush to get this merged? I'd appreciate it if we don't intentionally break stringdtype on main. I'm happy to help out with a fix.