@lysnikolaou 2023-09-22
> Could you add a release note?

Should we maybe add a release note when the other functions have been switched over to this implementation as well? Otherwise, should I just say `add` and `isalpha` have been re-implemented?