@mattip 2023-09-22
In the interest of limiting the scope of this first PR, to get it into the hands of users:

I think we can have a "fast path" for `np.bytes_` and `np.str_` only for now, and later rethink the `np.char` mechanism. That will not improve the performance for the new string dtype, but will also not break it. As @ngoldbaum says, there may be some overhaul that sees the common functions become full-fledged ufuncs.

As for the release note, I think it should simply state what has been done. If we go further we can edit the note or add a new one. There are over 60 functions in the `np.char` namespace, and not all of them are worthy of conversion.