@lysnikolaou 2023-09-22
> I think we can have a "fast path" for np.bytes_ and np.str_ only for now

Changed in https://github.com/numpy/numpy/pull/24756/commits/21f3a2ae46d2bfd5e075f569f0ba1a9f62b4b024.

I also added a release note and pushed the benchmarks I used, so this appears to be ready to be merged.