@ngoldbaum 2023-09-22
Cool, I can take a look at what's needed to bring these improvements to stringdtype once it's merged.