@lysnikolaou 2023-09-27
I've already got implementations ready for `find` and `rfind`. Should we merge this as-is or should I switch gears and look into making all of these into ufuncs?