@ngoldbaum 2023-09-27
If there's ongoing effort going into optimizing these and it's doable to make them ufuncs without too much extra pain, I would vote for making them ufuncs. It would be easier for me to support them in `stringdtype` if they were ufuncs. That's not a veto to continuing along this route, just me advocating for something that would make integrating these changes easier for me in the future.