@mattip 2023-09-27
Why would writing the inner loop for a ufunc be easier than writing the inner loop for the stringlib utility library? These feel like separate functions not amenable to ufunc generalization:
- in most cases the functions do not generalize to integer/float loops
- there is no need for axis handling nor need for a reduction form of the function
- providing an "out" kwarg will be complicated: what happens if the output width does not match? Error? Truncate the result? Teaching the behaviour of `out` could be difficult.